requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0,<2
matplotlib>=3.7.0
scikit-learn>=1.3.0
timm>=0.9.0
einops>=0.7.0
yacs>=0.1.8
pyyaml>=6.0
tqdm>=4.65.0
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
wandb>=0.16.0
scipy>=1.11.0
pandas>=2.0.0
Pillow>=10.0.0
